+++ src/kyoshin/help/CHelp_LearnArts.hpp
@@ -33,7 +33,7 @@ extern "C" u8* lbl_eu_806640F4;
 extern "C" UNKWORD func_80122450();
 extern "C" u8 func_8012246C();
 extern "C" u32 func_801B481C();
-extern "C" void* func_8009EC9C(u16 index);
+extern "C" void* func_8009EC9C(u32 index);
 extern "C" bool func_801F9268(unsigned char* p, int i, int j);
 
 // "wpn_type" BDAT column-name string (retail .rodata, defined in port/data_defs.cpp).
